Create Journeys Rescue
Alternatives to euthanasia for perfectly happy, healthy adoptable animals. The city shares the majority of responsibility with pet owners to keep and control  their pets/city for strays, abandoned, deceased, dumped and abused/surrendered or seized. People are not held accountable regardless of the meager laws set fourth for owning an animal and the responsibilities to be kept. The city is responsible for the shelter and feeding,,..etc. However just in my city alone, there are 20/40 animals euthanized every week or 2 due to over-population. They are not given enough time to be adopted or fostered for the amount of animals that are out there being thrown out, abused, tied up and left, strays in grocery store parking lots. et owners decide they dont  want their pet of 4years anymore etc. They are put down for the sole reason of not having room. The city does not honor its responsibilities with the serious nature it should  be nor do they enforce the laws on the owners or hold them accountable. So whos responsibility should it fall on. Rescues, Shelters and volunteers. not for profit org's. They are fighting all day every day to save, house, give or get medical care find fosters and adoptees at a rate too high to even mention. If the city cant keep up, and animals are being penalized and killed for the horrible abuse, mistreatment and blatent disregard for the safety or just plain responsibility as a human being. There is an answer. The city has numerous properties and buildings, etc...that are abandoned, seized or left from owners who have passed with no family to leave it to. DONATE some of them. There are plenty of poeple willing to help in so many different ways. It would solve overcrowding/population. It would save lives of amazing animals and give time to find good, responsible people. It would allow for volunteers to meet with possible adopter at their home. see how they interact, get along with pets, kids etc so that when they are placed with adopter, they have a 90% chance of not being back in a shelter awaiting death. People dont have to ove animals. they dont have to own or adopt. They just have to be human. Stop and gget that scared puppy off the hwy, out of a ditch or away from someone you know is hurting them. It might be esier for some people to walk away. I dont know how they can rest at night. Dont get close enough to take a picture, then go post and say you hope SOMEONE else will help. The donated properties would give people a place to take them instead of walking away because they are unable to take them in. To walk past because the laws are bigger on makinng someone who tries to pick up and save more responsible then the people that own/ed them. Theyre are people that attend meeting for the city trying to get laws changed. The property donations would be the start of keeping up with who has a pet. and keep records, check on animals to see if theyre still where people took them after deliberately going out looking to bring them home. who took the responsibility, on purpose. Give the animals a chance at a llififfefe of happiness and love that they deserve. And it can be done if given a place to have tht extra time to find the person to give them that. it wont solve everything but it will help rescues, shelters, volunteer vets and fosters to help so many more with the threat of euthansia falling in their hands alone. more space, more oppertunity to clean up the city so much more and put the responsibility where it belongs.